Documenting code
Optimised imports, and docstrings some functions, and reformatted some code
