+++ CODE/Logicytics.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
 
 import datetime
-import hashlib
-import shutil
 import threading
-from datetime import datetime
 from typing import Any
 
 from __lib_class import *
@@ -28,170 +25,9 @@ with many options and flags that can be used to customize its behavior.
 """
 
 
-class Zip:
-    """
-    A class to handle zipping files, generating SHA256 hashes, and moving files.
-
-    Methods:
-        __get_files_to_zip(path: str) -> list:
-            Returns a list of files to be zipped, excluding certain file types and names.
-
-        __create_zip_file(path: str, files: list, filename: str):
-            Creates a zip file from the given list of files.
-
-        __remove_files(path: str, files: list):
-            Removes the specified files from the given path.
-
-        __generate_sha256_hash(filename: str) -> str:
-            Generates a SHA256 hash for the specified zip file.
-
-        __write_hash_to_file(filename: str, sha256_hash: str):
-            Writes the SHA256 hash to a file.
-
-        __move_files(filename: str):
-            Moves the zip file and its hash file to designated directories.
-
-        and_hash(self, path: str, name: str, flag: str) -> tuple | str:
-            Zips files, generates a SHA256 hash, and moves the files.
-    """
-
-    @staticmethod
-    def __get_files_to_zip(path: str) -> list:
-        """
-        Returns a list of files and directories to be zipped, excluding certain file types and names.
-
-        Args:
-            path (str): The directory path to search for files.
-
-        Returns:
-            list: A list of file and directory names to be zipped.
-        """
-        excluded_extensions = (".py", ".exe", ".bat", ".ps1")
-        excluded_prefixes = ("config.", "SysInternal_Suite", "__pycache__")
-
-        return [
-            f for f in os.listdir(path)
-            if not f.endswith(excluded_extensions) and not f.startswith(excluded_prefixes)
-        ]
-
-    @staticmethod
-    def __create_zip_file(path: str, files: list, filename: str):
-        """
-        Creates a zip file from the given list of files.
-
-        Args:
-            path (str): The directory path containing the files.
-            files (list): A list of file names to be zipped.
-            filename (str): The name of the output zip file.
-
-        Returns:
-            None
-        """
-        with zipfile.ZipFile(f"{filename}.zip", "w") as zip_file:
-            for file in files:
-                if os.path.isdir(os.path.join(path, file)):
-                    for root, _, files in os.walk(os.path.join(path, file)):
-                        for f in files:
-                            zip_file.write(os.path.join(root, f), os.path.relpath(os.path.join(root, f), path))
-                else:
-                    zip_file.write(os.path.join(path, file))
-
-    @staticmethod
-    def __remove_files(path: str, files: list):
-        """
-        Removes the specified files from the given path.
-
-        Args:
-            path (str): The directory path containing the files.
-            files (list): A list of file names to be removed.
-
-        Returns:
-            None or str: Returns an error message if an exception occurs.
-        """
-        for file in files:
-            try:
-                shutil.rmtree(os.path.join(path, file))
-            except OSError:
-                os.remove(os.path.join(path, file))
-            except Exception as e:
-                return f"Error: {e}"
-
-    @staticmethod
-    def __generate_sha256_hash(filename: str) -> str:
-        """
-        Generates a SHA256 hash for the specified zip file.
-
-        Args:
-            filename (str): The name of the zip file.
-
-        Returns:
-            str: The SHA256 hash of the zip file.
-        """
-        with open(f"{filename}.zip", "rb") as zip_file:
-            zip_data = zip_file.read()
-        return hashlib.sha256(zip_data).hexdigest()
-
-    @staticmethod
-    def __write_hash_to_file(filename: str, sha256_hash: str):
-        """
-        Writes the SHA256 hash to a file.
-
-        Args:
-            filename (str): The name of the hash file.
-            sha256_hash (str): The SHA256 hash to be written.
-
-        Returns:
-            None
-        """
-        with open(f"{filename}.hash", "w") as hash_file:
-            hash_file.write(sha256_hash)
-
-    @staticmethod
-    def __move_files(filename: str):
-        """
-        Moves the zip file and its hash file to designated directories.
-
-        Args:
-            filename (str): The name of the files to be moved.
-
-        Returns:
-            None
-        """
-        shutil.move(f"{filename}.zip", "../ACCESS/DATA/Zip")
-        shutil.move(f"{filename}.hash", "../ACCESS/DATA/Hashes")
-
-    def and_hash(self, path: str, name: str, flag: str) -> tuple | str:
-        """
-        Zips files, generates a SHA256 hash, and moves the files.
-
-        Args:
-            path (str): The directory path containing the files.
-            name (str): The base name for the output files.
-            flag (str): A flag to be included in the output file names.
-
-        Returns:
-            tuple or str: A tuple containing success messages or an error message.
-        """
-        today = datetime.now()
-        filename = f"Logicytics_{name}_{flag}_{today.strftime('%Y-%m-%d_%H-%M-%S')}"
-        files_to_zip = self.__get_files_to_zip(path)
-        self.__create_zip_file(path, files_to_zip, filename)
-        check = self.__remove_files(path, files_to_zip)
-        if isinstance(check, str):
-            return check
-        else:
-            sha256_hash = self.__generate_sha256_hash(filename)
-            self.__write_hash_to_file(filename, sha256_hash)
-            self.__move_files(filename)
-            return (
-                f"Zip file moved to ../ACCESS/DATA/Zip/{filename}.zip",
-                f"SHA256 Hash file moved to ../ACCESS/DATA/Hashes/{filename}.hash",
-            )
-
-
 class Health:
     @staticmethod
-    def backup(directory: str, name: str) -> None:
+    def backup(directory: str, name: str):
         """
         Creates a backup of a specified directory by zipping its contents and moving it to a designated backup location.
 
@@ -243,7 +79,17 @@ class Health:
         return output, "info"
 
 
-def get_flags():
+def get_flags() -> tuple[str, str]:
+    """
+    Retrieves the command-line flags and sub-actions.
+
+    This function checks if the flags are provided as a tuple. If so, it attempts to unpack
+    the tuple into actions and sub_actions. If an exception occurs, it sets sub_actions to None.
+    If the flags are not a tuple, it prints the help message and exits the program.
+
+    Returns:
+        tuple: A tuple containing actions and sub_actions.
+    """
     if isinstance(Flag.data(), tuple):
         try:
             # Get flags
@@ -275,6 +121,14 @@ def special_execute(file_path: str):
 
 
 def handle_special_actions():
+    """
+    Handles special actions based on the provided action flag.
+
+    This function checks the value of the `action` variable and performs
+    corresponding special actions such as opening debug, developer, or extra
+    tools menus, updating the repository, restoring backups, creating backups,
+    or unzipping extra files.
+    """
     # Special actions -> Quit
     if action == "debug":
         log.info("Opening debug menu...")
@@ -333,6 +187,13 @@ def handle_special_actions():
 
 
 def check_privileges():
+    """
+    Checks if the script is running with administrative privileges and handles UAC (User Account Control) settings.
+
+    This function verifies if the script has admin privileges. If not, it either logs a warning (in debug mode) or
+    prompts the user to run the script with admin privileges and exits. It also checks if UAC is enabled and logs
+    warnings accordingly.
+    """
     if not Check.admin():
         if DEBUG == "DEBUG":
             log.warning("Running in debug mode, continuing without admin privileges - This may cause issues")
@@ -343,8 +204,7 @@ def check_privileges():
             exit(1)
 
     if Check.uac():
-        log.warning("UAC is enabled, this may cause issues")
-        log.warning("Please disable UAC if possible")
+        log.warning("UAC is enabled, this may cause issues - Please disable UAC if possible")
 
 
 def generate_execution_list(actions: str) -> list | list[str] | list[str | Any]:
@@ -472,28 +332,28 @@ def execute_scripts():
 
 def zip_generated_files():
     """Zips generated files based on the action."""
-    if action == "modded":
-        zip_loc_mod, hash_loc = Zip().and_hash("..\\MODS", "MODS", action)
-        log.info(zip_loc_mod)
-        zip_values = Zip().and_hash("..\\MODS", "MODS", action)
+
+    def zip_and_log(directory, name):
+        zip_values = FileManagement.Zip.and_hash(directory, name, action)
         if isinstance(zip_values, str):
             log.error(zip_values)
         else:
-            zip_loc_mod, hash_loc = zip_values
-            log.info(zip_loc_mod)
+            zip_loc, hash_loc = zip_values
+            log.info(zip_loc)
             log.debug(hash_loc)
 
-    zip_values = Zip().and_hash(".", "CODE", action)
-    if isinstance(zip_values, str):
-        # If error, log it
-        log.error(zip_values)
-    else:
-        zip_loc, hash_loc = zip_values
-        log.info(zip_loc)
-        log.debug(hash_loc)
+    if action == "modded":
+        zip_and_log("..\\MODS", "MODS")
+    zip_and_log(".", "CODE")
 
 
 def handle_sub_action():
+    """
+    Handles sub-actions based on the provided sub_action flag.
+
+    This function checks the value of the `sub_action` variable and performs
+    corresponding sub-actions such as shutting down or rebooting the system.
+    """
     if sub_action == "shutdown":
         subprocess.call("shutdown /s /t 3", shell=False)
     elif sub_action == "reboot":
